Kaubad

+++ index.php
@@ -15,7 +15,7 @@ if(isset($_REQUEST["sort"])) {
 if(isset($_REQUEST["search_term"])) {
     $search_term = $_REQUEST["search_term"];
 }
-if(isset($_REQUEST["kaubagrupp_lisamine"])) {
+if(isset($_REQUEST["kaubagrupp_lisamine"])&& isAdmin()) {
     addGroup($_REQUEST["kaubagrupp"]);
     header("Location: index.php");
     exit();
@@ -25,7 +25,7 @@ if(isset($_REQUEST["kauba_lisamine"])) {
     header("Location: index.php");
     exit();
 }
-if(isset($_REQUEST["delete"])) {
+if(isset($_REQUEST["delete"])&& isAdmin()) {
     deleteItem($_REQUEST["delete"]);
 }
 if(isset($_REQUEST["save"])) {
@@ -42,12 +42,12 @@ $goods = groupData($sort, $search_term);
 </head>
 <body>
 <header class="header">
-    <p>Admin on sisse logitud</p>
+    <p class="text"><?=$_SESSION["kasutaja"]?> on sisse logitud</p>
     <form action="logout.php" method="post">
-        <input type="submit" value="Logi välja" name="logout">
+        <input type="submit" value="Logi välja" name="logout" class="sub">
     </form>
     <div class="container">
-        <h1>Tabelid | Kaubad ja kaubagruppid</h1>
+        <h1 class="text">Tabelid | Kaubad ja kaubagruppid</h1>
     </div>
 </header>
 <main class="main">
@@ -76,10 +76,10 @@ $goods = groupData($sort, $search_term);
         <table>
             <thead>
             <tr>
-                <th><a href="index.php?sort=Id">Id</a></th>
-                <th><a href="index.php?sort=kaubanimi">Kaubanimi</a></th>
-                <th><a href="index.php?sort=hind">Hind</a></th>
-                <th><a href="index.php?sort=kaubagrupp">Kaubagrupp</a></th>
+                <th><a href="index.php?sort=Id" class="text">Id</a></th>
+                <th><a href="index.php?sort=kaubanimi" class="text">Kaubanimi</a></th>
+                <th><a href="index.php?sort=hind" class="text">Hind</a></th>
+                <th><a href="index.php?sort=kaubagrupp" class="text">Kaubagrupp</a></th>
                 <th></th>
             </tr>
             </thead>
@@ -91,34 +91,38 @@ $goods = groupData($sort, $search_term);
                     <td><?=$product->hind ?></td>
                     <td><?=$product->kaubagrupp ?></td>
                     <td>
+                        <?php if(isAdmin()){?>
                         <a title="Kustuta inimene" class="deleteBtn" href="index.php?delete=<?=$product->id?>"
                            onclick="return confirm('Oled kindel, et soovid kustutada?');">X</a>
                         <a title="Muuda inimest" class="editBtn" href="index.php?edit=<?=$product->id?>">&#9998;</a>
+                        <?php } ?>
                     </td>
                 </tr>
             <?php endforeach; ?>
             </tbody>
         </table>
+        <?php if(isAdmin()){?>
         <form action="index.php">
-            <h2>Kaubagrupi lisamine:</h2>
+            <h2 class="text">Kaubagrupi lisamine:</h2>
             <dl>
-                <dt>Kaubagrupp nimi:</dt>
+                <dt class="text">Kaubagrupp nimi:</dt>
                 <dd><input type="text" name="kaubagrupp" placeholder="Sisesta nimi..."></dd>
-                <input type="submit" name="kaubagrupp_lisamine" value="Lisa kaubagrupp">
+                <input type="submit" name="kaubagrupp_lisamine" value="Lisa kaubagrupp" class="sub">
             </dl>
         </form>
+        <?php } ?>
         <form action="index.php">
-            <h2>Kauba lisamine:</h2>
+            <h2 class="text">Kauba lisamine:</h2>
             <dl>
-                <dt>Kaubanimi:</dt>
+                <dt class="text">Kaubanimi:</dt>
                 <dd><input type="text" name="kaubanimi" placeholder="Sisesta kaubanimi..."></dd>
-                <dt>Hind:</dt>
+                <dt class="text">Hind:</dt>
                 <dd><input type="text" name="hind" placeholder="Sisesta hind..."></dd>
-                <dt>Kaubagrupp:</dt>
+                <dt class="text">Kaubagrupp:</dt>
                 <dd><?php
                     echo createSelect("SELECT id, kaubagrupp FROM kaubagrupid", "kaubagrupp_id");
                     ?></dd>
-                <input type="submit" name="kauba_lisamine" value="Lisa kauba">
+                <input type="submit" name="kauba_lisamine" value="Lisa kauba" class="sub">
             </dl>
         </form>
     </div>
